Partly covered ArgAdapter with unit tests
- Fixed bug in ArgAdapter.setArgValue()
- Moved and refactored addContent logic from FullElement to FullElementContentType
- AstWomTestFixture.parse* -> AstWomTestFixture.postprocess*

+++ swc-engine/src/main/java/org/sweble/wikitext/engine/astwom/adapters/ArgAdapter.java
@@ -171,7 +171,8 @@ public class ArgAdapter
 		
 		Toolbox.expectType(ValueAdapter.class, value);
 		
-		this.value.unlink();
+		if (this.value != null)
+			this.value.unlink();
 		
 		this.value = (ValueAdapter) value;
 		this.value.link(this, this.name, null);
